moved role selection component to auth dir
