+++ frontend/src/App.js
@@ -13,7 +13,7 @@ import { Box, CircularProgress } from '@mui/material'
 import RequireAuth from './components/Auth/RequireAuth'
 import RequireProfile from './components/Auth/RequireProfile'
 import RequireAuthAndNoProfile from './components/Auth/RequireAuthAndNoProfile.js'
-import RoleSelection from './components/RoleSelection'
+import RoleSelection from './components/Auth/RoleSelection.js'
 import StudentProfileForm from './components/StudentProfileForm'
 import FacultyProfileForm from './components/FacultyProfileForm'
 import InvalidEmail from './components/Auth/InvalidEmail'
