+++ frontend/src/__tests__/Auth/RoleSelection.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
-import RoleSelection from '../../components/RoleSelection'
+import RoleSelection from '../../components/Auth/RoleSelection'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
 import '@testing-library/jest-dom'
 import { MemoryRouter } from 'react-router-dom'
